Consolidate overlay plotting into neuron_export
